requirements.txt
requests
python-socketio
appdirs
pyperclip